+++ src/index.ts
@@ -13,14 +13,16 @@ export { default as useSort } from './hooks/useSort';
 export * from './hooks/callbacks';
 
 // Components
+export { default as Breadcrumbs } from './components/Breadcrumbs';
 export { default as CioPlp, CioPlpProvider } from './components/CioPlp';
 export { default as CioPlpGrid, ZeroResults } from './components/CioPlpGrid';
-export { default as ProductCard } from './components/ProductCard';
-export { default as Sort } from './components/Sort';
-export { default as Pagination } from './components/Pagination';
 export { default as Filters } from './components/Filters';
 export { default as Groups } from './components/Groups';
+export { default as MobileModal } from './components/MobileModal';
+export { default as Pagination } from './components/Pagination';
+export { default as ProductCard } from './components/ProductCard';
 export { default as ProductSwatch } from './components/ProductSwatch';
+export { default as Sort } from './components/Sort';
 export { default as Spinner } from './components/Spinner';
 
 // Utils
